Handle case of no documentation changes

+++ tools/ci/publish_docs.sh
@@ -62,21 +62,26 @@ cd output/$GH_REPO_NAME
 # both exist. This is a good indication that Doxygen did it's work.
 if [ -d "docs/head" ] && [ -f "docs/head/index.html" ]; then
 
-    echo 'Uploading documentation to the gh-pages branch...'
     # Add everything in this directory (the Doxygen code documentation) to the
     # gh-pages branch.
     # GitHub is smart enough to know which files have changed and which files have
     # stayed the same and will only update the changed files.
     git add --all
 
-    # Commit the added files with a title and description containing the Travis CI
-    # build number and the GitHub commit reference that issued this build.
-    git commit -m "Deploy code docs to GitHub Pages Travis build: ${TRAVIS_BUILD_NUMBER}" -m "Commit: ${TRAVIS_COMMIT}"
+    # Check to see if there are any differences in the documentation.
+    if ! git diff-index --quiet HEAD; then
+      echo 'Uploading documentation to the gh-pages branch...'
+      # Commit the added files with a title and description containing the Travis CI
+      # build number and the GitHub commit reference that issued this build.
+      git commit -m "Deploy code docs to GitHub Pages Travis build: ${TRAVIS_BUILD_NUMBER}" -m "Commit: ${TRAVIS_COMMIT}"
 
-    # Force push to the remote gh-pages branch.
-    # The ouput is redirected to /dev/null to hide any sensitive credential data
-    # that might otherwise be exposed.
-    git push --force "https://${GH_REPO_TOKEN}@${GH_REPO_REF}" > /dev/null 2>&1
+      # Force push to the remote gh-pages branch.
+      # The ouput is redirected to /dev/null to hide any sensitive credential data
+      # that might otherwise be exposed.
+      git push --force "https://${GH_REPO_TOKEN}@${GH_REPO_REF}" > /dev/null 2>&1
+    else
+      echo 'No documentation changes. Doing nothing.'
+    fi
 else
     echo '' >&2
     echo 'Warning: No documentation (html) files have been found!' >&2
